+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [0-based versioning](https://0ver.org/).
 
 ## [Unreleased]
 
+### Added
+
+- Watchdog example
+
 ## v0.14.2 - 13-September-2025
 
 ### Changed